+++ hyperdbg/libhyperdbg/code/app/libhyperdbg.cpp
@@ -878,7 +878,10 @@ HyperDbgLoadVmmModule()
     }
     else
     {
-        ShowMessages("vmx operation is not supported by your processor\n");
+#ifdef ENV_WINDOWS
+        ShowMessages("vmx operation is not supported by your processor "
+                     "(if you are using an Intel processor, it might be because VBS is not disabled!)\n");
+#endif
         return 1;
     }